Change RowsetWriter num_rows() return int64_t (#1535)

+++ be/src/olap/rowset/rowset_writer.h
@@ -55,7 +55,7 @@ public:
 
     virtual Version version() = 0;
 
-    virtual int32_t num_rows() = 0;
+    virtual int64_t num_rows() = 0;
 
     virtual RowsetId rowset_id() = 0;
